Update clinicaltrials_download to use tframe with NULL query
I was getting the “No query parameters given” error when trying to use
a tframe to download CT.gov results. The first if statement in the code
triggers the error because the tframe argument isn’t taken into
account. I restructured the main if-else block to build the appropriate
query given either a query or tframe. Tested successfully with a tframe
and query, both with a NULL count and count = 90.