clinicaltrials_download retrieves >100 tframe results
Updated clinicaltrials_download to retrieve all tframe results,
regardless of length. final_url changed to a vector of final_urls,
which contains a list of 100-count urls to retrieve. This list is
looped through to download and extract multiple zip files.

This is important to minimize the result set if one wants to filter a
clinicaltrial_search result set by the status of the study, and the
resulting list has >100 nct_ids.

The extraction step is pretty slow for >100 XML files, so there’s a
warning that the command could take a while. A progress bar for the
extraction step would be a nice touch, possibly using the pbapply
package.