Fix error for missing org_study_id, verbose downloads, parallel processing
Some XML files from NCT.gov are missing the org_study_id field, and
this causes an error in the parse_study_xml function. I used tryCatch
to trap this error.

I also added a verbose option to clinicaltrials_download, which shows
the number of downloaded documents and the file name for each file
being processed by parse_study_xml. This assists with trapping errors
in XML parsing.

I also added a check for the parallel package, and use it if
appropriate in running parse_study_xml on the downloaded XML files. In
testing, this cut processing time in half for n = 915.